+++ retz-server/src/main/java/io/github/retz/db/Database.java
@@ -454,15 +454,16 @@ public class Database {
         }
     }
 
-    public List<Job> listJobs(String id, Job.JobState state, Optional<String> tag) throws SQLException {
+    public List<Job> listJobs(String id, Job.JobState state, Optional<String> tag, int limit) throws SQLException {
         List<Job> ret = new LinkedList<>();
         String prefix = "SELECT j.json FROM jobs j, applications a WHERE j.appid = a.appid AND a.owner = ?";
-        String sql = prefix + " AND j.state=?";
+        String sql = prefix + " AND j.state=? ORDER BY j.id ASC LIMIT ?";
 
         try (Connection conn = dataSource.getConnection(); // pool.getConnection();
              PreparedStatement p = conn.prepareStatement(sql)) {
             p.setString(1, id);
             p.setString(2, state.toString());
+            p.setInt(3, limit);
 
             conn.setAutoCommit(true);